+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Function Pigeon::assertNothingDispatched().
+- Function Pigeon::assertDispatchedCount(int $count).
 ## [v2.1.0]
 ### Added
 - Laravel 9 support